# document/swagger.yaml
openapi: 3.0.0
info:
  title: Charum
  description: |-
    This is a Charum Back End
  version: 1.0.0
servers:
  - url: https://charum-api.nyakit.in/api/v1
...
paths:
...
  /user/profile:
    get:
      tags:
        - User
      summary: Get user profile
      description: Get user's profile data
      operationId: GetProfile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success to login
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2Mzc5ZmU4ZGFkMzNlYzQyNTE3MDBkZWQiLCJyb2xlIjoidXNlciIsImlzcyI6ImNoYXJ1bSIsImV4cCI6MTY2ODk0Nzg3NiwibmJmIjoxNjY4OTQ0Mjc2LCJpYXQiOjE2Njg5NDQyNzZ9._hk2Y2pt6tADTN0aHTukPNd2O_wRdHsgn3Vo4jIX_zc
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: validation failed
                  data:
                    type: object
                    properties:
                      field:
                        type: string
                        example: password
                      message:
                        type: string
                        example: This field must be at least 8 characters
                  pagination:
                    type: string
                    example: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: email is not registered
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
    put:
      tags:
        - User
      summary: Update user profile
      description: Update user's profile data
      operationId: UserUpdate
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@charum.com
                userName:
                  type: string
                  example: charum-user
                displayName:
                  type: string
                  example: charum user
                biodata:
                  type: string
                  example: halo aku user
                socialMedia:
                  type: string
                  example: instagram @anon
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success to update
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: invalid id
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: unauthorized
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: failed to get user
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 409
                  message:
                    type: string
                    example: email is already used
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: failed to update user
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
...
  /admin/user/id/{id}:
    get:
      tags:
        - Admin
      summary: Get user profile
      description: Get user's profile data
      operationId: GetUserByID
      parameters:
        - in: path
          name: id
          description: user ID
          required: true
          schema:
            type: string
            example: 637dfc9bb09d0bdd7a2fe3de
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success to get user by id
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: invalid id
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null

        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: failed to get user
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
    put:
      tags:
        - Admin
      summary: Update user profile
      description: Update user's profile data
      operationId: AdminUpdate
      parameters:
        - in: path
          name: id
          description: user ID
          required: true
          schema:
            type: string
            example: 637dfc9bb09d0bdd7a2fe3de
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@charum.com
                userName:
                  type: string
                  example: charum-user
                displayName:
                  type: string
                  example: charum user
                biodata:
                  type: string
                  example: halo aku user
                socialMedia:
                  type: string
                  example: instagram @anon
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success to update
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: invalid id
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: unauthorized
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: failed to get user
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 409
                  message:
                    type: string
                    example: email is already used
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: failed to update user
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
    delete:
      tags:
        - Admin
      summary: Delete user
      description: Delete user's data
      operationId: Delete
      parameters:
        - in: path
          name: id
          description: user ID
          required: true
          schema:
            type: string
            example: 637dfc9bb09d0bdd7a2fe3de
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success to delete user
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: invalid id
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: unauthorized
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: failed to get user
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: failed to delete user
                  data:
                    type: string
                    example: null
                  pagination:
                    type: string
                    example: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: 6379fe8dad33ec4251700ded
        email:
          type: string
          example: user@charum.com
        userName:
          type: string
          example: charumuser
        displayName:
          type: string
          example: charum user
        password:
          type: string
          example: secretpassword
        isActive:
          type: boolean
          example: true
        role:
          type: string
          example: user
        createdAt:
          type: string
          example: 2022-11-26T23:05:11.806+07:00
        updatedAt:
          type: string
          example: 2022-11-26T23:05:11.806+07:00
...
security:
  - bearerAuth: []